Use sys.executable for spawning Python subprocesses
The original code assumes that the Python interpreter used for running
pDynamo scripts is called "python". This is not necessarily true on
systems with multiple Python installations. On many modern
installations, "python" is Python 3, whereas Python 2 has some other
name.

With this patch, examples and tutorials are run in the same Python
interpreter that was specified on the command line when running
RunExamples or RunTutorials.